+++ types/core/api/AviationWeatherAeroflyApi.d.ts
@@ -2,5 +2,6 @@ import { type AeroflyFlight } from "@fboes/aerofly-custom-missions";
 import { AviationWeatherApi } from "./AviationWeatherApi.js";
 export declare class AviationWeatherApiAerofly extends AviationWeatherApi {
     fetchMetarToFlight(airportCode: string, flight: AeroflyFlight): Promise<AeroflyFlight>;
+    fetchTafToFlight(airportCode: string, flight: AeroflyFlight): Promise<AeroflyFlight>;
 }
 //# sourceMappingURL=AviationWeatherAeroflyApi.d.ts.map